rename and move AssayTypeStatusGrid.tsx
